+++ src/UglyToad.PdfPig/Parser/FileStructure/CrossReferenceOffsetValidator.cs
@@ -12,9 +12,9 @@
             this.offsetValidator = offsetValidator;
         }
 
-        public long Validate(long crossReferenceOffset, ISeekableTokenScanner scanner, IInputBytes bytes, bool isLenientParsing)
+        public int Validate(int crossReferenceOffset, ISeekableTokenScanner scanner, IInputBytes bytes, bool isLenientParsing)
         {
-            long fixedOffset = offsetValidator.CheckXRefOffset(crossReferenceOffset, scanner, bytes, isLenientParsing);
+            int fixedOffset = offsetValidator.CheckXRefOffset(crossReferenceOffset, scanner, bytes, isLenientParsing);
             if (fixedOffset > -1)
             {
                 crossReferenceOffset = fixedOffset;
